Use i32 instead of i64 and compile primitives

+++ src/vm/instr.rs
@@ -3,7 +3,7 @@ pub enum Instruction {
     PushUnit,
     Pop,
 
-    PushInt(i64),
+    PushInt(i32),
     AddInt,
     SubInt,
     MulInt,
